+++ src/application/use-cases/employee/delete-employee-use-case.ts
@@ -1,10 +1,12 @@
 import { Employee } from "../../../core/entities/employee";
 import { NotFoundError } from "../errors/not-found-error";
+import { UnauthorizedError } from "../errors/unauthorized-error";
 import { EmployeeRepository } from "../../../adapters/repositories/employee-repository";
 import { Either, failure, success } from "../../../utils/either";
 
 type DeleteEmployeeUseCaseRequest = {
   id: string;
+  id_employee: string;
 };
 
 class DeleteEmployeeUseCase {
@@ -12,11 +14,16 @@ class DeleteEmployeeUseCase {
 
   public async execute({
     id,
-  }: DeleteEmployeeUseCaseRequest): Promise<Either<NotFoundError, Employee>> {
+    id_employee,
+  }: DeleteEmployeeUseCaseRequest): Promise<Either<NotFoundError | UnauthorizedError, Employee>> {
     const employeeExists = await this.employeeRepository.findById(id);
 
     if (!employeeExists) {
       return failure(new NotFoundError(`Nenhum funcionário encontrado com o ID: ${id}`));
+    }
+
+    if (id !== id_employee) {
+      return failure(new UnauthorizedError(`Você não tem permissão para excluir esse funcionário`));
     }
 
     const employee = await this.employeeRepository.delete(id);
